force usage of HybridCluster branded loader menu
orbbw was hijacked in hc source tree to be a hybrid cluster logo.
For 9.2 the upstream has changed a default logo to a tribute one that is
branded with Nakatomi BSD 9.2 theme as can be seen in Die Hard movie.
So now we have to explicit about which logo we actually.

Long term, it's better to restore orbbw to its original look and re-add
the hc logo as a new log and then the default to it.

logo