hc-install: refine the logic for building a list of available disks
Explicitly skip a disk if diskinfo(8) fails on it.
That means that the disk drive has no media or is not usable in some
other way.

Also, change sorting to a more natural one.
Previous sorting was useful when we had duplicate device entries for
the same disk: adX vs adaX.

While there, squash a few bash-isms that did not provide any added value
comparing to posix shell equivalents.