Revert "hc-install: do not wipe all disks prematurely"
This reverts commit 4c27f05f93d7b0dc55b5bc1a358eded84ce123f8.
Unfortunately with the current state of the matters in the installer we
have to wipe out all the disks to keep things simple.
Otherwise, for example, if some disks were a part of hpool but are not
selected for new hpool, then zfs won't allow us to either destroy the
old pool or create the new one with same name.

Probably we should have two separate disk selection dialogs: one for
selecting disks to destroy and the other to select disks for a new pool.

Perhaps we should add special logic to find all disks comprising
existing hpool and destroy them.